backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-allauth==0.57.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
django-crispy-forms==2.1
crispy-bootstrap5==0.7
Pillow==10.1.0
requests==2.31.0
python-decouple==3.8
django-cacheops==7.2
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
scikit-learn==1.3.2
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
requests==2.31.0
pandas==2.1.4
numpy==1.25.2
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
